Bulk upload work in progress
Bulk upload is sitll a bit buggy where sometimes it would upload to the db and sometimes it would not.

Adde signalR to perform like percentage tracking

added in a bulk upload log,

Bulk upload is only working for Users other implementations do not work and need to be fixed.

SignalR still buggy and does not display the migration progress